moved s3 upload function into a controller file
